Fix GCC build warnings
There should be no functional changes. Only cleaner code and build log.

+++ lib/src/rpmsg_lib/pru_virtqueue.c
@@ -98,7 +98,7 @@ int16_t pru_virtqueue_add_used_buf (
     num = vq->vring.num;
     used = vq->vring.used;
 
-    if (head > num)
+    if (head > (int)num)
         return PRU_VIRTQUEUE_INVALID_HEAD;
 
     /*
